Expand Parameters to include mutants
Source: R/util_model.R

Expand Parameters to include mutants. All mutants get the same schedule, equal to all the unique times that any resident was introduced (if mutants) or the default schedule (if residents). This results in more work than is really needed, but should be reasonable most of the time.

expand_parameters(trait_matrix, p, mutant = TRUE)

Arguments

trait_matrix
A matrix of traits corresponding to the new types to introduce.

p
A Parameters object.

mutant
Create new types as mutants? These will have no effect on other plants in the community (i.e. have zero density).