Reference Growth Model
Source: R/reference_plant.R

Create an instance of reference version of growth model FF16

make_reference_plant_FF16()

Value

An object with a number of methods (accessed with $). These will hopefully be documented at some point.

Details

This is an old version of the growth model, implemented by Daniel Falster. It does not implement any of the actual growth. It is used in testing code to make sure that the full model version gives sensible results.